task2 (CNN-LSTM)/requirements.txt
numpy==1.19.5
tensorflow==2.9.1
keras==2.9.0
pandas==1.3.5
pillow==9.2.0
nltk==3.7
scikit-learn==1.0.2
matplotlib==3.5.3
opencv-python==4.6.0.66
tqdm==4.64.1 